requirements.txt
# Slack Bot
slack-bolt==1.18.0
slack-sdk==3.23.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Configuration
python-dotenv==1.0.0

# Dashboard (optional)
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
